Remove unused functionalities from internal xcodeproj.
