--- internal/xcodeproj/code_sign_info_script.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-
-# test code_sign_info_script.rb
-
-export user=""
-export project=""
-export scheme=""
-export configuration=""
-
-ruby code_sign_info_script.rb
