+++ internal/xcodeproj/project.go
@@ -1,10 +1,13 @@
 package xcodeproj
 
 import (
+	"bufio"
 	"fmt"
 	"path/filepath"
+	"regexp"
 	"strings"
 
+	"github.com/bitrise-io/go-utils/fileutil"
 	"github.com/bitrise-io/go-utils/pathutil"
 )
 
@@ -33,7 +36,7 @@ func NewProject(xcodeprojPth string) (ProjectModel, error) {
 		return ProjectModel{}, fmt.Errorf("Project descriptor not found at: %s", pbxprojPth)
 	}
 
-	sdks, err := GetBuildConfigSDKs(pbxprojPth)
+	sdks, err := getBuildConfigSDKs(pbxprojPth)
 	if err != nil {
 		return ProjectModel{}, err
 	}
@@ -41,7 +44,7 @@ func NewProject(xcodeprojPth string) (ProjectModel, error) {
 	project.SDKs = sdks
 
 	// Shared Schemes
-	schemes, err := ProjectSharedSchemes(xcodeprojPth)
+	schemes, err := projectSharedSchemes(xcodeprojPth)
 	if err != nil {
 		return ProjectModel{}, err
 	}
@@ -49,7 +52,7 @@ func NewProject(xcodeprojPth string) (ProjectModel, error) {
 	project.SharedSchemes = schemes
 
 	// Targets
-	targets, err := ProjectTargets(xcodeprojPth)
+	targets, err := projectTargets(xcodeprojPth)
 	if err != nil {
 		return ProjectModel{}, err
 	}
@@ -59,12 +62,56 @@ func NewProject(xcodeprojPth string) (ProjectModel, error) {
 	return project, nil
 }
 
-// ContainsSDK ...
-func (p ProjectModel) ContainsSDK(sdk string) bool {
-	for _, s := range p.SDKs {
-		if s == sdk {
-			return true
+func getBuildConfigSDKs(pbxprojPth string) ([]string, error) {
+	content, err := fileutil.ReadStringFromFile(pbxprojPth)
+	if err != nil {
+		return []string{}, err
+	}
+
+	return getBuildConfigSDKsFromContent(content)
+}
+
+func getBuildConfigSDKsFromContent(pbxprojContent string) ([]string, error) {
+	sdkMap := map[string]bool{}
+
+	beginXCBuildConfigurationSection := `/* Begin XCBuildConfiguration section */`
+	endXCBuildConfigurationSection := `/* End XCBuildConfiguration section */`
+	isXCBuildConfigurationSection := false
+
+	// SDKROOT = macosx;
+	pattern := `SDKROOT = (?P<sdk>.*);`
+	regexp := regexp.MustCompile(pattern)
+
+	scanner := bufio.NewScanner(strings.NewReader(pbxprojContent))
+	for scanner.Scan() {
+		line := scanner.Text()
+
+		if strings.TrimSpace(line) == endXCBuildConfigurationSection {
+			break
+		}
+
+		if strings.TrimSpace(line) == beginXCBuildConfigurationSection {
+			isXCBuildConfigurationSection = true
+			continue
+		}
+
+		if !isXCBuildConfigurationSection {
+			continue
+		}
+
+		if match := regexp.FindStringSubmatch(line); len(match) == 2 {
+			sdk := match[1]
+			sdkMap[sdk] = true
 		}
 	}
-	return false
+	if err := scanner.Err(); err != nil {
+		return []string{}, err
+	}
+
+	sdks := []string{}
+	for sdk := range sdkMap {
+		sdks = append(sdks, sdk)
+	}
+
+	return sdks, nil
 }
